server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
bs4==0.0.1
requests==2.26.0
python-dotenv==1.0.1